feat(shared): add infisical secret delete
